requirements.txt
absl-py==2.3.1
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
cssselect==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
fake-useragent==2.2.0
feedparser==6.0.11
filelock==3.20.0
finnhub-python==2.4.25
flatbuffers==25.2.10
fonttools==4.59.1
frozendict==2.4.6
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.2
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multitasking==0.0.12
namex==0.1.0
networkx==3.5
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pandas-datareader==0.10.0
parse==1.20.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.32.0
pycparser==2.22
pyee==11.1.1
Pygments==2.19.2
pyparsing==3.2.3
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-html==0.10.0
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yahoo-fin==0.8.9.1
yfinance==0.2.58
zipp==3.23.0
